More gracefully handle file includes
Instead of mindlessly doing "cwd + $tp4_string" and hoping for the best,
strip the trailing slashes from "cwd"; attempt to include from "cwd" and
if that fails, fallback to the folder from where we're executing (as is
done by the PHP-based parser).

Also, when it fails, we now provide a nice, human-readable, error
message.